refactor: ProductListLayout 타입 추가 #7
- OpenMarketViewController의 Layout 역할을 분리함
